requirements.txt
psycopg2-binary>=2.9.10
python-dotenv>=1.0.0
requests>=2.28.0
tqdm>=4.64.0
pytest>=7.0.0
pytest-cov>=4.0.0
pandas>=1.5.0
numpy>=1.23.0


